Update tests to use pytest
