+++ .isort.cfg
@@ -1,8 +1,6 @@
 [settings]
 default_section=THIRDPARTY
 indent='    '
-known_yunojuno=yunojuno
-known_tests=yunojuno.tests
 sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,TESTS,LOCALFOLDER
 multi_line_output=3
 line_length=88
