+++ tox.ini
@@ -10,14 +10,14 @@ python =
 
 [testenv]
 deps =
-    coverage
+    pytest
+    pytest-django
+    pytest-cov
     django22: Django>=2.2,<2.3
     django30: Django>=3.0,<3.1
 
 commands =
-    coverage erase
-    coverage run --branch  --include=side_effects/* runtests.py
-    coverage report -m
+    pytest --ds=tests.settings --cov=side_effects --verbose tests
 
 [testenv:lint]
 deps =
